refactor: centralize testing setup for PrelaunchSignUpForm in test-utils
- Moved setup function for PrelaunchSignUpForm tests to new file under test-utils/components/. This will allow multiple test suites to use the same setup.
- Added JSDoc comment to setup function for better clarity and maintainability.
